reorganization, pre-playtest updates
