+++ Assets/Materials/Study/HandScannerCorrect.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3c165641048655d4983b6c969f11a4b2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
