+++ Assets/Materials/Study/HandScannerIncorrect.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6296d0395629bf24492deefc84462487
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
