+++ Assets/Scripts/Common/NPCAnimator.cs.meta
