+++ Assets/Scripts/Puzzles/2BioWall/BioWallTablet.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8a64055ce1ac42148b9d71690cf2ed97
+guid: 5932c0b99720c0e4d9be23b394a1adeb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
