+++ Assets/Scripts/Puzzles/4Piano/CartridgeDisc.cs
