+++ Assets/Scripts/Puzzles/4Piano/CartridgeSlot.cs
