+++ Assets/Scripts/Puzzles/HandScannerLever/SecretDeskButton.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5932c0b99720c0e4d9be23b394a1adeb
+guid: 2a7635d4faa76ca478db2dace69b8db9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
